+++ lib/wrappers/title.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 import 'package:objd/basic/widget.dart';
 import 'package:objd/basic/command.dart';
 import 'package:objd/basic/widgets.dart';
@@ -5,7 +7,7 @@ import 'package:objd/build/build.dart';
 import 'package:objd/basic/entity.dart';
 import 'package:objd/basic/text_components.dart';
 import 'package:meta/meta.dart';
-import 'package:gson/gson.dart';
+
 
 class Title extends RestActionAble {
   String type;
@@ -15,17 +17,17 @@ class Title extends RestActionAble {
   Title(Entity selector, {@required List<TextComponent> show}){
     type = "title";
     entity = selector.toString();
-    jsonText = gson.encode(show.map((text) => text.toMap()).toList());
+    jsonText = json.encode(show.map((text) => text.toMap()).toList());
   }
   Title.subtitle(Entity selector, {@required List<TextComponent> show}){
     type = "subtitle";
     entity = selector.toString();
-    jsonText = gson.encode(show.map((text) => text.toMap()).toList());
+    jsonText = json.encode(show.map((text) => text.toMap()).toList());
   }
   Title.actionbar(Entity selector, {@required List<TextComponent> show}){
     type = "actionbar";
     entity = selector.toString();
-    jsonText = gson.encode(show.map((text) => text.toMap()).toList());
+    jsonText = json.encode(show.map((text) => text.toMap()).toList());
   }
   /// Title.clear clears all titles again:
   Title.clear(Entity selector){